// helpers/types.ts
export enum eTelosNetwork{
  telosTestnet = "telosTestnet",
  telos = "telos"
}

export enum EthereumNetworkNames {
  kovan = "kovan",
  ropsten = "ropsten",
  main = "main",
  matic = "matic",
  mumbai = "mumbai",
  xdai = "xdai",
  avalanche = "avalanche",
  fuji = "fuji",
}

export enum AavePools {
  proto = "proto",
  matic = "matic",
  amm = "amm",
  avalanche = "avalanche",
}

export enum eContractid {
  Example = "Example",
  PoolAddressesProvider = "PoolAddressesProvider",
  MintableERC20 = "MintableERC20",
  MintableDelegationERC20 = "MintableDelegationERC20",
  PoolAddressesProviderRegistry = "PoolAddressesProviderRegistry",
  PoolConfigurator = "PoolConfigurator",
  ValidationLogic = "ValidationLogic",
  ReserveLogic = "ReserveLogic",
  GenericLogic = "GenericLogic",
  Pool = "Pool",
  PriceOracle = "PriceOracle",
  Proxy = "Proxy",
  MockAggregator = "MockAggregator",
  AaveOracle = "AaveOracle",
  DefaultReserveInterestRateStrategy = "DefaultReserveInterestRateStrategy",
  LendingPoolCollateralManager = "LendingPoolCollateralManager",
  InitializableAdminUpgradeabilityProxy = "InitializableAdminUpgradeabilityProxy",
  MockFlashLoanReceiver = "MockFlashLoanReceiver",
  WalletBalanceProvider = "WalletBalanceProvider",
  AToken = "AToken",
  MockAToken = "MockAToken",
  DelegationAwareAToken = "DelegationAwareAToken",
  MockStableDebtToken = "MockStableDebtToken",
  MockVariableDebtToken = "MockVariableDebtToken",
  AaveProtocolDataProvider = "AaveProtocolDataProvider",
  IERC20Detailed = "IERC20Detailed",
  StableDebtToken = "StableDebtToken",
  VariableDebtToken = "VariableDebtToken",
  FeeProvider = "FeeProvider",
  TokenDistributor = "TokenDistributor",
  StableAndVariableTokensHelper = "StableAndVariableTokensHelper",
  ATokensAndRatesHelper = "ATokensAndRatesHelper",
  UiPoolDataProviderV3 = "UiPoolDataProviderV3",
  WrappedTokenGatewayV3 = "WrappedTokenGatewayV3",
  WETH = "WETH",
}
// ...
export type iAssetsWithoutUSD<T> = Omit<iAssetBase<T>, "USD">;

export type iMosaicPoolAssets<T> = Pick<
  iAssetsWithoutUSD<T>,
  "DAI" | "USDC" | "WBTC" | "WETH"
>;

export type iMantlePoolAssets<T> = Pick<
  iAssetsWithoutUSD<T>,
  "DAI" | "USDC" | "USDT" | "WBTC" | "WETH" 
>;

export type iTelosPoolAssets<T> = Pick<
  iAssetsWithoutUSD<T>,
  "TLOS" | "USDC" | "USDT" | "BTCb" | "ETH"
>;

export type iMultiPoolsAssets<T> = iAssetCommon<T> | iMosaicPoolAssets<T>;

export type iAavePoolTokens<T> = Omit<iMosaicPoolAssets<T>, "ETH">;

export type iAssetAggregatorBase<T> = iAssetsWithoutETH<T>;

export enum TokenContractId {
  DAI = "DAI",
  WETH = "WETH",
  USDC = "USDC",
  USDT = "USDT",
  WBTC = "WBTC",
}

export interface IReserveParams
  extends IReserveBorrowParams,
    IReserveCollateralParams {
  aTokenImpl: eContractid;
  reserveFactor: string;
  supplyCap: string;
  strategy: IInterestRateStrategyParams;
}

export interface IStrategy {
  [key: string]: IInterestRateStrategyParams;
}

export interface IInterestRateStrategyParams {
  name: string;
  optimalUsageRatio: string;
  baseVariableBorrowRate: string;
  variableRateSlope1: string;
  variableRateSlope2: string;
  stableRateSlope1: string;
  stableRateSlope2: string;
  baseStableRateOffset: string;
  stableRateExcessOffset: string;
  optimalStableToTotalDebtRatio: string;
}

export interface IReserveBorrowParams {
  borrowingEnabled: boolean;
  stableBorrowRateEnabled: boolean;
  reserveDecimals: string;
  borrowCap: string;
  debtCeiling: string;
  borrowableIsolation: boolean;
  flashLoanEnabled: boolean;
}

export interface IReserveCollateralParams {
  baseLTVAsCollateral: string;
  liquidationThreshold: string;
  liquidationBonus: string;
  liquidationProtocolFee?: string;
}
// ...
export interface IBaseConfiguration {
  SynthexAddress?: tEthereumAddress;
  MarketId: string;
  ATokenNamePrefix: string;
  StableDebtTokenNamePrefix: string;
  VariableDebtTokenNamePrefix: string;
  SymbolPrefix: string;
  ProviderId: number;
  PythId?: string;
  TestnetMarket?: boolean;
  ProviderRegistryOwner?: iParamsPerNetwork<tEthereumAddress | undefined>;

  FallbackOracle?: OracleConfig;
  PriceOracle: OracleConfig;

  WrappedTokenGateway?: iParamsPerNetwork<tEthereumAddress>;
  ReserveFactorTreasuryAddress: iParamsPerNetwork<tEthereumAddress>;
  StableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  VariableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  ReserveAssets?: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  
  ReservesConfig: SymbolMap<IReserveParams>;
  WrappedNativeTokenSymbol: string; // ex: WETH or WMATIC
  IncentivesConfig: IncentivesConfig;
  EModes: SymbolMap<EMode>;
  L2PoolEnabled?: iParamsPerNetwork<boolean>;
  StkAaveProxy?: iParamsPerNetwork<tEthereumAddress>;
  ParaswapRegistry?: iParamsPerNetwork<tEthereumAddress>;
  FlashLoanPremiums: {
    total: number;
    protocol: number;
  };
  RateStrategies: IStrategy;
}

export interface ICommonConfiguration extends IBaseConfiguration {}

export interface IMosaicConfiguration extends ICommonConfiguration {
  // ReservesConfig: iMosaicPoolAssets<IReserveParams>;
}
export interface ITelosConfiguration extends ICommonConfiguration {
  ReservesConfig: iTelosPoolAssets<IReserveParams>;
}